requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
lxml==5.1.0
pycryptodome==3.20.0
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
urllib3==1.26.5
